Port some more metrics to MetricsContainer. (#1934)
* Port some more metrics to MetricsContainer.
* chore: Update jicoco.
* fix: VideobridgeMetricsContainer static instance returns correct type.

+++ jvb/src/main/kotlin/org/jitsi/videobridge/Endpoint.kt
@@ -1001,7 +1001,7 @@ class Endpoint @JvmOverloads constructor(
 
         conference.videobridge.statistics.apply {
             val bweStats = transceiverStats.bandwidthEstimatorStats
-            bweStats.getNumber("incomingEstimateExpirations")?.toInt()?.let {
+            bweStats.getNumber("incomingEstimateExpirations")?.toLong()?.let {
                 incomingBitrateExpirations.addAndGet(it)
             }
             totalKeyframesReceived.addAndGet(transceiverStats.rtpReceiverStats.videoParserStats.numKeyframes)
